requirements.txt
Django>=4.2.0,<4.3
django-filter==23.2
django-cors-headers==3.14
djangorestframework>=3.14.0,<3.15
psycopg2-binary==2.9.6
Pillow==9.5.0

openpyxl==3.1.2
lxml==4.9.3

reportlab==4.0.4
qrcode==7.4.2

djangorestframework-simplejwt[crypto]

drf-yasg # swagger
requests
ipython
PyJWT==2.8.0
